|| Dev branch ||
Work-in-progress implementation of Pixy2 communication over I2C. Writing to the Pixy seems to be working, but reading from it is inconsistent at best.

Changed RunningAverageFilter to SlidingAverageFilter to more accurately reflect what it actually does. Additionally, changed the implementation of the filter - instead of a queue-like array, the filter now uses a circular buffer to avoid large array copy operations.

Added TrapezoidalMotionProfile. This class takes a list of (x,y) coordinates (nodes) and constructs a profile that linearly interpolates between adjacent nodes.

Experimental implementation of a Velocity PID in VIPD. P and D terms are not currently working.

Added CubicBezierCurve. This class represents a Bezier curve in 2D space. Aggressively optimized to increase the performance of the findPointClosestTo() method within the class.

Added GuidingVectorField (work-in-progress). This class provides guidance vectors that converge to a provided path as well as converging to the end of the path. While it currently only supports a single
CubicBezierCurve as a path, I plan to add support for composite Bezier curves.

New vector-based version of MecanumEncoderCalculator. This new version is easier to understand than the previous version. Said previous version has been renamed ScalarMecanumEncoderCalculator, and will likely be deprecated or removed.

Added Fusion2. This class is a navigation system that uses only the drivetrain encoders to navigate in autonomous. It is technically a form of dead reckoning, but it's quite accurate. In testing with our current robot,
Fusion2 can drive multiple meters to a target and stop within about 3cm of said target. At the cost of movement speed, this accuracy can be improved to within 5mm.

Added new operations to Vector2:
 - dot: dot product of 2 vectors.
 - crossMag: crossMag is the magnitude of the cross product of 2 vectors, if we pretend the 2D vectors are actually 3D vectors with a z-component of 0.
 - normalized: normalizes a vector. this retains the heading, but scales the vector so it has a magnitude of 1.
Additionally, add() and subtract() now support varargs, allowing for more concise chains of vector operations.

Added OdometerNavSource. This is a work-in-progress implementation of 2-wheel odometry. Its kinematics were derived by using vector calculus to find the position of the robot by modeling the motion of the odometer through field-relative space.
Currently it supports 0th- and 1st-order models of robot motion. I have worked out an equation for a 2nd-order model, but it uses the Fresnel integrals. I'm still in the process of implementing (or finding) a method to approximate these integrals.