Initial implementation of PathPlanner, GuidingVectorField, and parametric curves
MecanumDrive: during the SkyStone season, I had to invert all of the drivetrain motors quickly. I did this in MecanumDrive, then forgot to remove it. It has been fixed now, but every movement axis has been inverted because of it.

Removed some old/unused code.

Added parametric curves. Currently the only supported curves are line segments, cubic Bezier curves, and arcs.

Changed PIDF's deltaTime calculation to use nanoseconds instead of milliseconds. This should make the D term a little more stable.

Added Vector3. Vector3 is basically the 3D form of Vector2.

Added OdometerKinematics. This class contains a generalized form of odometer kinematics, which is used by both TwoWheelOdometerNavSource and ThreeWheelOdometerNavSource.

MathUtilities: work-in-progress support for Fresnel integrals. It's not working yet.

Added TwoWheelOdometerNavSource (work-in-progress). This NavSource allows field-relative localization with 2 odometer wheels and the IMU.

Added ThreeWheelOdometerNavSource. This NavSource allows field-relative localization with 3 odometer wheels.

Added PathPlanner (work-in-progress). This class not only allows the robot to follow pre-programmed paths, but it also self-correcting. That is, it will automatically move back to the path if pushed off of it.