Updated to SDK v9.1
The new SDK version brings a lot of changes to the SDK's support for computer vision. This broke a lot of old code (namely the VuMark-based navigation system and TFDetector). That old code has been removed, as
it has been made obsolete by developments in odometry and changes to the SDK.