wip fix of Pixy2
ThorCore has had code for I2C support of the Pixy2 for quite a while. however, due to a REV firmware bug, work on the driver was abandoned in favor of USB support. The newest version of the SDK includes an I2C driver for the HuskyLens, and the comments in that driver describe the exact behavior of the bug. With that information in hand, the Pixy2 I2C driver can finally be fixed.